src/test/java/org/bexterlab/tetrisbackend/core/StartGameInteractorImplTest.java
package org.bexterlab.tetrisbackend.core;


import org.bexterlab.tetrisbackend.controller.StartGameInteractor;
import org.bexterlab.tetrisbackend.core.exception.InvalidUsernameException;
import org.bexterlab.tetrisbackend.core.exception.YouAlreadyHaveAGameException;
import org.bexterlab.tetrisbackend.core.mock.GameStoreFake;
import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;

class StartGameInteractorImplTest {

    private GameStoreFake gameStore;
    StartGameInteractor startGameInteractor;

    @BeforeEach
    void setUp() {
        gameStore = new GameStoreFake();
        startGameInteractor = new StartGameInteractorImpl(gameStore);
    }

    @Test
    public void invalidUsernameTest() {
        Assertions.assertThrows(InvalidUsernameException.class, () -> startGameInteractor.start("%start_with_Special"));
        Assertions.assertThrows(InvalidUsernameException.class, () -> startGameInteractor.start("short"));
        Assertions.assertThrows(InvalidUsernameException.class, () -> startGameInteractor.start("greater_then_20_chara"));
        Assertions.assertThrows(InvalidUsernameException.class, () -> startGameInteractor.start("test._"));
        Assertions.assertThrows(InvalidUsernameException.class, () -> startGameInteractor.start("test_."));
    }

    @Test
    public void hasAGameTest() {
        gameStore.hasGameWithUser = true;
        Assertions.assertThrows(YouAlreadyHaveAGameException.class, () -> startGameInteractor.start("valid_user"));
    }

    @Test
    public void successTestTest() {
        gameStore.hasGameWithUser = false;
        Assertions.assertEquals("test", startGameInteractor.start("valid_user"));
    }
}